tests/wpt/web-platform-tests/compat/css-style-declaration-alias-enumeration.html

<!DOCTYPE html>
<title>Aliases enumerated on CSSStyleDeclaration</title>
<link rel="help" href="https://compat.spec.whatwg.org/#css-simple-aliases">
<meta name="assert" content="This test verifies that aliases are enumerated on CSSStyleDeclaration" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
</body>
<script>
const PREFIXED_PROPS = [
  'webkitAlignContent',
  'webkitAlignItems',
  'webkitAnimationName',
  'webkitAlignSelf',
  'webkitAnimationDuration',
  'webkitAnimationTimingFunction',
  'webkitAnimationIterationCount',
  'webkitAnimationDirection',
  'webkitAnimationPlayState',
  'webkitAnimationDelay',
  'webkitAnimationFillMode',
  'webkitAnimation',
  'webkitBackfaceVisibility',
  'webkitBackgroundOrigin',
  'webkitBackgroundSize',
  'webkitBorderBottomLeftRadius',
  'webkitBorderBottomRightRadius',
  'webkitBorderTopLeftRadius',
  'webkitBorderTopRightRadius',
  'webkitBorderRadius',
  'webkitBoxShadow',
  'webkitFlex',
  'webkitFlexBasis',
  'webkitFlexDirection',
  'webkitFlexFlow',
  'webkitFlexGrow',
  'webkitFlexShrink',
  'webkitFlexWrap',
  'webkitFilter',
  'webkitJustifyContent',
  'webkitMask',
  'webkitMaskBoxImage',
  'webkitMaskBoxImageOutset',
  'webkitMaskBoxImageRepeat',
  'webkitMaskBoxImageSlice',
  'webkitMaskBoxImageSource',
  'webkitMaskBoxImageWidth',
  'webkitMaskClip',
  'webkitMaskComposite',
  'webkitMaskImage',
  'webkitMaskOrigin',
  'webkitMaskPosition',
  'webkitMaskRepeat',
  'webkitMaskSize',
  'webkitOrder',
  'webkitPerspective',
  'webkitPerspectiveOrigin',
  'webkitTransformOrigin',
  'webkitTransformStyle',
  'webkitTransform',
  'webkitTransitionDelay',
  'webkitTransitionDuration',
  'webkitTransitionProperty',
  'webkitTransitionTimingFunction',
  'webkitTransition',
];

const docBodyStyle = document.body.style;
for (let prop of PREFIXED_PROPS) {
  test(() => {
    assert_true(prop in docBodyStyle);
  }, `${prop} found on CSSStyleDeclaration`);
}
</script>
